+++ soul/backends/__init__.py
@@ -1,3 +1,4 @@
-from ._file import FileBackend
+﻿from ._file import FileBackend
+from ._sqlite import SqliteBackend
 
-__all__ = ["FileBackend"]
+__all__ = ["FileBackend", "SqliteBackend"]